Solved problem with Server 500 errors.
Needed to ensure proper fixtures were being loaded. This should be
addressed by naming the files properly, but for the time being, just
added them to the load-data script that I use to deploy to staging and
production. Also updated some files for readability, added hosts to
the ALLOWED_HOSTS list for the staging server, and updated the
Supervisor configuration file to set the DJANGO_SETTINGS_MODULE
environment properly. This seems better than relying on the
postactivate script in the virtual environment, which is hard to place
under revision control.